+++ src/unbuf.rs
@@ -194,6 +194,12 @@ impl TlsServer {
         Ok(Self { sc })
     }
 
+    /// Get immutable access to the wrapped
+    /// `UnbufferedServerConnection`, if available
+    pub fn connection(&self) -> Option<&UnbufferedServerConnection> {
+        self.sc.as_ref()
+    }
+
     /// Process as much data as possible, moving data between `ext`
     /// and `int`.  `ext` is the pipe which typically carries TLS
     /// protocol data to/from an external TCP connection.  `int` is
@@ -260,6 +266,12 @@ impl TlsClient {
         };
 
         Ok(Self { cc })
+    }
+
+    /// Get immutable access to the wrapped
+    /// `UnbufferedClientConnection`, if available
+    pub fn connection(&self) -> Option<&UnbufferedClientConnection> {
+        self.cc.as_ref()
     }
 
     /// Process as much data as possible, moving data between `ext`