+++ CHANGELOG.md
@@ -7,6 +7,14 @@ Rustls has breaking changes.
 
 <!-- see keepachangelog.com for format ideas -->
 
+## 0.23.1 (2024-09-16)
+
+### Added
+
+- `TlsClient::connection` and `TlsServer::connection` to get immutable
+  access to the wrapped Rustls structure, for obtaining `server_name`,
+  ALPN info, or whatever.
+
 ## 0.23.0 (2024-04-14)
 
 ### Added